pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "meshtastic-mqtt-monitor"
version = "0.1.0"
description = "A Python tool for monitoring and decoding Meshtastic MQTT traffic"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "GPL-3.0-or-later"}
authors = [
    {name = "Meshtastic MQTT Monitor Contributors"}
]
keywords = ["meshtastic", "mqtt", "monitor", "protobuf", "mesh-network"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications",
    "Topic :: System :: Monitoring",
]

dependencies = [
    "paho-mqtt>=1.6.0,<3.0.0",
    "meshtastic>=2.0.0,<3.0.0",
    "cryptography>=41.0.0,<43.0.0",
    "PyYAML>=6.0,<7.0",
    "colorama>=0.4.6,<1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0,<9.0.0",
    "pytest-mock>=3.10.0,<4.0.0",
    "pytest-cov>=4.0.0,<6.0.0",
    "black>=23.0.0,<25.0.0",
    "flake8>=6.0.0,<8.0.0",
    "mypy>=1.0.0,<2.0.0",
]

[project.scripts]
meshtastic-monitor = "main:main"

[project.urls]
Homepage = "https://github.com/meshtastic/meshtastic-mqtt-monitor"
Repository = "https://github.com/meshtastic/meshtastic-mqtt-monitor"
Issues = "https://github.com/meshtastic/meshtastic-mqtt-monitor/issues"

[tool.setuptools]
py-modules = ["main"]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.yml"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=src --cov-report=term-missing"
